+++ example/lib/states/settings_states.dart
@@ -1,0 +1,6 @@
+// Theme state
+class ThemeState {
+  final bool isDark;
+
+  const ThemeState({this.isDark = false});
+}