+++ multi_state_example/android/app/src/main/kotlin/com/example/multi_state_example/MainActivity.kt
@@ -1,5 +1,0 @@
-package com.example.multi_state_example
-
-import io.flutter.embedding.android.FlutterActivity
-
-class MainActivity : FlutterActivity()